Fixed problems for missing components and json without a version number
Yida ran into a problem with some requests that didn't have a version
number so I fixed that and I also made a catch-all for any components
which are not in the maven_p2 database. Prior to this it would not
return anything for an item not found. It sets state of the component
that is not found to unavailable.